utilityArmAngleMath

+++ src/org/usfirst/frc/team1619/robot2016/IO/SensorInput.java
@@ -69,7 +69,7 @@ public class SensorInput {
     robotOut.setDriveRightPos(position);
   }
   
-  public int getDartPosition() {
+  public double getDartPosition() {
     return robotOut.getDartPosition() / Constants.DART_ENC_TICKS_PER_INCH;
   }
